feat:añadiendo auths

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/app/auth/actions/auth-actions'
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup'
@@ -32,11 +33,15 @@ const postSchema = yup.object({
 
 
 export async function POST(request: Request) {
+
+    const user = await getUserSessionServer()
+    if(!user) return NextResponse.json('No autorizado',{status:401})
+
     try {
 
         const { complete, description } = await postSchema.validate(await request.json())
 
-        const todo = await prisma.todo.create({ data: { complete, description } })
+        const todo = await prisma.todo.create({ data: { complete, description,day:'x', userId:user.id } })
 
         return NextResponse.json(todo)
     } catch (error) {
@@ -46,11 +51,12 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+
+    const user = await getUserSessionServer()
+    if(!user) return NextResponse.json('No autorizado',{status:401})
+
     try {
-
-
-        const todo = await prisma.todo.deleteMany({ where: { complete: true } })
-
+        const todo = await prisma.todo.deleteMany({ where: { complete: true,userId:user.id } })
         return NextResponse.json('Exito de borrado')
     } catch (error) {
         return NextResponse.json(error, { status: 400 })
